models/malaria_model/requirements.txt
torch
torchvision
flask
flask-cors
pillow
numpy
opencv-python
efficientnet-pytorch
